2. Add Lease 3. Add Expense 4. Get all Lease 5. Get all Expenses

+++ property_manager_be/src/main/java/com/property_manager/property_manager_be/repository/LeaseRepository.java
@@ -3,4 +3,4 @@ package com.property_manager.property_manager_be.repository;
 import com.property_manager.property_manager_be.models.Lease;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LeaseRepository extends CrudRepository<Lease, Integer> {}
+public interface LeaseRepository extends CrudRepository<Lease, Long> {}
