+++ property_manager_be/src/main/java/com/property_manager/property_manager_be/repository/PropertyRepository.java
@@ -3,4 +3,4 @@ package com.property_manager.property_manager_be.repository;
 import com.property_manager.property_manager_be.models.Property;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PropertyRepository extends CrudRepository<Property, Integer> {}
+public interface PropertyRepository extends CrudRepository<Property, Long> {}
